Refactored birch_girder into multiple modules
This takes the monolithic __init__.py and refactors it into multiple modules
based on the event type.

This also pulls out common functions into utils.py and moves templates into
templates.py.

Logging is broken out into setup_logging.py and the Alerter class is removed.
Instead of the Alerter.alert function, the logging setup has been enriched
such that all critical logs are emitted to SNS as the alert function did.

The EventHandler class is also removed and instead data is passed to functions
as arguments (e.g. config and event). persistent_data is now also passed as
an argument instead of a global in the child modules.

To support this new model with multiple modules, everything is moved from the
root of the AWS Lambda function directory into a birch_girder directory. The
plugins are also moved under this new directory.

This still leaves the process_email.py module as too large but I'll consider
breaking that down further in a later commit.

To see the previous structure before refactoring, you'll need to go to the
previous commit, 34768a6150dad63f5c68fdab84ed999594a0713a

config.example.yaml is updated to include any missing config options.
